Remove type interdependencies between modules
