+++ packages/core/src/index.tsx
@@ -37,11 +37,12 @@ import {
     DATADOG_GRAPH_QL_VARIABLES_HEADER
 } from './rum/instrumentation/resourceTracking/graphql/graphqlHeaders';
 import type { FirstPartyHost } from './rum/types';
-import { ErrorSource, PropagatorType, RumActionType } from './rum/types';
+import { PropagatorType, RumActionType } from './rum/types';
 import { DatadogProvider } from './sdk/DatadogProvider/DatadogProvider';
 import { DdSdk } from './sdk/DdSdk';
 import { FileBasedConfiguration } from './sdk/FileBasedConfiguration/FileBasedConfiguration';
 import { DdTrace } from './trace/DdTrace';
+import { ErrorSource } from './types';
 import { DefaultTimeProvider } from './utils/time-provider/DefaultTimeProvider';
 import type { Timestamp } from './utils/time-provider/TimeProvider';
 import { TimeProvider } from './utils/time-provider/TimeProvider';
