+++ packages/core/src/logs/DdLogs.ts
@@ -4,18 +4,17 @@
  * Copyright 2016-Present Datadog, Inc.
  */
 
+import { DdAttributes } from '../DdAttributes';
 import { DATADOG_MESSAGE_PREFIX, InternalLog } from '../InternalLog';
 import { SdkVerbosity } from '../SdkVerbosity';
 import type { DdNativeLogsType } from '../nativeModulesTypes';
-import { DdAttributes } from '../rum/DdAttributes';
-import type { ErrorSource } from '../rum/types';
+import type { ErrorSource, LogEventMapper } from '../types';
 import { validateContext } from '../utils/argsUtils';
 
 import { generateEventMapper } from './eventMapper';
 import type {
     DdLogsType,
     LogArguments,
-    LogEventMapper,
     LogWithErrorArguments,
     NativeLogWithError,
     RawLogWithError
